pyproject.toml
[tool.poetry]
name = "heigit-disaster-portal"
version = "0.1.0"
description = "Disaster Area Portal on basic FastAPI infrastructure using Docker, Gunicorn, FastAPI and Python 3.8 on alpine linux"
authors = ["Amandus Butzer <amandus.butzer@heigit.org>", "Takara Baumbach <takara.baumbach@heigit.org"]

[tool.poetry.dependencies]
alembic = "^1.7.7"  # database migration tool
python = "^3.8"
python-dotenv = "^0.20.0"  # reading .env files
fastapi = "^0.75.2"  # api framework
email-validator = "^1.1.1"  # pydantic email validation in api docs
python-multipart = "^0.0.5"  # pydantic file upload and form data in api docs
geojson-pydantic = "^0.3.3"  # pydantic geojson validation
SQLAlchemy = "^1.4.35"  # orm database interaction
GeoAlchemy2 = "^0.11.1"  # handling geometries with sqlalchemy
aiofiles = "^0.8.0 "  # file responses and serving static files
databases = "^0.5.3[postgresql,sqlite]"  # async support for PostgreSQL and SQLite
psycopg2-binary = "^2.9.2"  # PostgreSQL adapter
tenacity = "^8.0.1"  # retrying library (waiting for database to restart)
uvicorn = "^0.17.6"  # ASGI server
geopy = "^2.1.0"  # geographic calculation functions
python-dateutil = "^2.8.1"  # correct timestamp parsing
requests = "^2.27.1"  # easy HTTP request interaction
PyYAML = "^6.0"  # YAML support

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"  # testing
pytest-cov = "^3.0.0"  # test coverage
pytest-asyncio = "^0.18.3"  # async testing
httpx = "^0.22.0"  # needed for async testing
black = "^22.3.0"  # TODO: used ?
isort = "^5.10.1"  # TODO: used ?
autoflake = "^1.3.1"  # TODO: used ?
mypy = "^0.942"  # TODO: used ?

[build-system]
requires = ["poetry>=1.1.0b2"]
build-backend = "poetry.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"  # recommended unless multiple async frameworks are used
minversion = "7.0"
testpaths = [
    "tests"
]
log_auto_indent = "True"  # auto-indent multilines
